requirements.txt
absl-py==1.3.0
anyio==3.6.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
branca==0.6.0
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
configobj==5.0.6
contourpy==1.0.5
customtkinter==4.6.3
cycler==0.10.0
darkdetect==0.7.1
decorator==5.1.1
dulwich==0.20.50
everett==3.1.0
exceptiongroup==1.0.4
executing==1.1.1
fastapi==0.88.0
folium==0.13.0
fonttools==4.38.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
glob2==0.7
google-auth==2.13.0
google-auth-oauthlib==0.4.6
googlemaps==4.6.0
gpxpy==1.5.0
grpcio==1.50.0
h11==0.14.0
idna==2.10
imutils==0.5.4
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pymavlink==2.4.37
pyparsing==2.4.7
pyperclip==1.8.2
pyproj==3.4.0
pyrsistent==0.19.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
selenium==4.6.0
semantic-version==2.10.0
sentry-sdk==1.11.1
Shapely==1.8.5.post1
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.22.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tkhtmlview==0.1.1.post5
tkintermapview==1.18
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
wsproto==1.2.0
wurlitzer==3.0.3
zmq==0.0.0
